refactor: move map in a dedicated component
