+++ src/pages/map.tsx
@@ -1,12 +1,11 @@
-import { useEffect, useRef, useState } from 'react';
-import { Geolocation } from '@capacitor/geolocation';
+import { useState } from 'react';
 import { Select } from '../components/select';
 import { Tag } from '../components/tag';
 import { Bars2Icon } from '@heroicons/react/24/outline';
 import { MapPinIcon } from '@heroicons/react/24/solid';
 import Navigation from '../components/navigation';
 import store from '../store';
-import mapboxgl from '../services/mapbox';
+import MapContainer from '../components/map-container';
 
 const MapView = () => {
   // TODO: take in from db
@@ -35,43 +34,10 @@ const MapView = () => {
     setIsOpen(!isOpen);
   };
 
-  // Mapbox
-  const mapContainer = useRef(null);
-  const map = useRef<mapboxgl.Map | null>(null);
-  const [lng, setLng] = useState(0);
-  const [lat, setLat] = useState(0);
-  const [zoom] = useState(12);
-
-  // User location
-  const getCurrentPosition = async () => {
-    const pos = await Geolocation.getCurrentPosition();
-    setLat(pos.coords.latitude);
-    setLng(pos.coords.longitude);
-    map.current?.setCenter([pos.coords.longitude, pos.coords.latitude]);
-  };
-
-  useEffect(() => {
-    if (map.current) return;
-
-    map.current = new mapboxgl.Map({
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      container: mapContainer.current!,
-      style: 'mapbox://styles/mapbox/streets-v11',
-      center: [lng, lat],
-      zoom: zoom,
-    });
-
-    map.current?.on('style.load', () => {
-      map.current?.setFog({}); // Set the default atmosphere style
-    });
-
-    getCurrentPosition();
-  });
-
   return (
     <>
       <div className="absolute inset-0 h-full w-full">
-        <div ref={mapContainer} className="map-container h-full w-full"></div>
+        <MapContainer />
       </div>
       <Navigation />
       <div className="relative w-full">
